refactor(macos): split the 2,942-line platform/macos.rs by concern
platform/macos.rs had grown into a grab-bag of unrelated macOS concerns.
Break it into focused submodules behind the same public path (macos.rs is now
just ResolvedTheme + re-exports), so each file is one concern:

- appearance: theme resolution + panel appearance/material
- window:     spotlight/HUD/sticky panels, geometry, resize, WebKit tuning,
              and the native Show-More bar (driven by the window resize path)
- icon:       app-icon extraction (.icns fast path + NSWorkspace fallback)
- input:      global key monitors (Cmd-Q, snippet expansion) + press-and-hold
- system:     accessibility, frontmost app, running-app activation

Pure code moves — the existing tests move next to their code and all pass.
